Intro/Positioning
I filled in the first 2 sections. Please feel free to edit anything I
have there. I left the abbreviations section with N/A but if you think
of any that need to be added feel free to update it. I’ll check back in
a day or two to add more to whatever else needs done.